Refactorored scrollfix to hook
