+++ packages/next-dashboard/sass/components/_modal.scss
@@ -1,7 +1,3 @@
-.dashboard_modal-active {
-  position: fixed;
-}
-
 .modal-overlay {
   display: flex;
   justify-content: flex-start;
